### 2025-04-13 dfd6c9a
Updated DOIs for JOSS paper
