+++ paper.bib
@@ -1,13 +1,3 @@
-
-@misc{fidgit,
-  author = {A. M. Smith and K. Thaney and M. Hahnel},
-  title = {Fidgit: An ungodly union of GitHub and Figshare},
-  year = {2020},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/arfon/fidgit}
-}
-
 @article{BakerDenoising,
    author = {Baker, Rebecca R. and Muthurangu, Vivek and Rega, Marilena and Walsh, Stephen B. and Steeden, Jennifer A.},
    title = {Rapid 2D 23Na MRI of the calf using a denoising convolutional neural network},
@@ -15,7 +5,7 @@
    volume = {110},
    pages = {184-194},
    ISSN = {0730-725X},
-   DOI = {https://doi.org/10.1016/j.mri.2024.04.027},
+   DOI = {10.1016/j.mri.2024.04.027},
    url = {https://www.sciencedirect.com/science/article/pii/S0730725X24001401},
    year = {2024},
    type = {Journal Article}
@@ -29,6 +19,7 @@
    number = {5},
    pages = {C479-C504},
    ISSN = {1064-8275},
+   DOI = {10.1137/18m120885x}
    year = {2019},
    type = {Journal Article}
 }
@@ -41,7 +32,7 @@
    number = {6},
    pages = {1768-1776},
    ISSN = {0740-3194},
-   DOI = {https://doi.org/10.1002/mrm.24389},
+   DOI = {10.1002/mrm.24389},
    url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/mrm.24389},
    year = {2013},
    type = {Journal Article}
@@ -55,7 +46,7 @@
    number = {1},
    pages = {266-279},
    ISSN = {0740-3194},
-   DOI = {https://doi.org/10.1002/mrm.29855},
+   DOI = {10.1002/mrm.29855},
    url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/mrm.29855},
    year = {2024},
    type = {Journal Article}
@@ -82,6 +73,7 @@
    publisher = {IEEE},
    pages = {688-697},
    ISBN = {1665435771},
+   DOI = {10.1109/IPDPSW52791.2021.00105},
    year = {2021},
    type = {Conference Proceedings}
 }
