+++ paper.bib
@@ -19,7 +19,7 @@
    number = {5},
    pages = {C479-C504},
    ISSN = {1064-8275},
-   DOI = {10.1137/18m120885x}
+   DOI = {10.1137/18m120885x},
    year = {2019},
    type = {Journal Article}
 }
