### 2025-04-29 #8000 @Kevin-Mattheus-Moerman
@jennifersteeden Dear author, thanks for this submission. I am the AEiC on this track and here to help process the initial steps. I will also be the handling editor. Before we proceed, please can you have a look at the following points:

- [ ] Although your project has a basic statement on receiving contributions we recommend that you add more detailed contributing guidelines. Please work to add these and to mention/link to them in your README. You could for instance create a `CONTRIBUTING.md` file (see here for some examples: https://contributing.md/example/).
