+++ actions-runner/externals/node20/include/node/node_version.h
@@ -23,7 +23,7 @@
 #define SRC_NODE_VERSION_H_
 
 #define NODE_MAJOR_VERSION 20
-#define NODE_MINOR_VERSION 18
+#define NODE_MINOR_VERSION 19
 #define NODE_PATCH_VERSION 0
 
 #define NODE_VERSION_IS_LTS 1
