### 2025-05-19 #8000 @jennifersteeden
@Kevin-Mattheus-Moerman I think it is because I need to increase the LFS storage on GitHub. Until I have managed to sort this with my institute then I cannot make any changes. The JOSS paper seems to be linked to v0.23.1 and not the master branch - so although I have made the updates to the references, they are not reflected in the JOSS paper which the review process is referencing (the master branch is correct). I will get around to this as my priority this week. Kind regards, Jennifer Steeden
