+++ actions-runner/_work/tensorflow-mri/tensorflow-mri/paper.md
@@ -1,91 +1,0 @@
----
-title: 'TensorFlow MRI: An Open-Source TensorFlow Library for MRI Applications'
-tags:
-  - Magnetic Resonance Imaging
-  - MRI
-  - Medical Imaging
-  - Reconstruction
-  - Post-processing
-  - Segmentation
-  - Machine Learning
-  - Python
-authors:
-  - name: Javier Montalt-Tordera
-    orcid: 0000-0003-1956-1409
-    affiliation: 1
-  - name: Vivek Muthurangu
-    orcid: 0000-0002-4292-6456
-    affiliation: 1
-  - name: Olivier Jaubert
-    orcid: 0000-0002-7854-4150
-    affiliation: 1
-  - name: Rebecca Rose Baker
-    orcid: 0000-0002-5016-7912
-    affiliation: "1, 2"
-  - name: Tina Yao
-    orcid: 0000-0003-0470-311X
-    affiliation: "1, 3"
-  - name: Michele Pascale
-    orcid: 0000-0001-6853-2685
-    affiliation: 1
-  - name: Christopher Watson
-    affiliation: 1
-  - name: Anirudh Raman
-    orcid: 0009-0005-8301-0742
-    affiliation: 1
-  - name: Ruaraidh Campbell
-    orcid: 0000-0002-7457-9065
-    affiliation: "1, 3"
-  - name: Jennifer Steeden
-    orcid: 0000-0002-9792-2022
-    corresponding: true # (This is how to denote the corresponding author)
-    affiliation: 1
-affiliations:
- - name: UCL Centre for Translational Cardiovascular Imaging, University College London, Zayed Centre for Research, London, UK.
-   index: 1
-   ror: 02jx3x895   
- - name: UCL Centre for Medical Imaging, University College London, London, UK.
-   index: 2
- - name: UK Research and Innovation Centre for Doctoral Training in AI-enabled Healthcare Systems, University College London, London, UK.
-   index: 3
-date: 26 February 2025
-bibliography: paper.bib
----
-
-# Summary
-
-Open-source reconstruction software is vital for the ongoing development of accelerated magnetic resonance imaging (MRI). Current state-of-the-art techniques are complex and often time-consuming, which hinders uptake of accelerated techniques in the clinical environment. Recently, deep learning (DL) has revolutionized MRI reconstruction, reducing reconstruction time while maintaining high image quality. However, deployment in the research and clinical environment requires a of combination of modern DL frameworks, such as TensorFlow, with MRI reconstruction operations. Thus, we present *‘TensorFlow MRI’*, an open-source library of TensorFlow operators for rapid computational MRI, particularly focussed on reconstruction (DL and non-DL) and post-processing for MRI data. 
-
-# Statement of need
-
-*‘TensorFlow MRI’* is a library of TensorFlow operators for computational magnetic resonance imaging. Its purpose is to simplify the creation of end-to-end MRI reconstruction and processing pipelines, within a unified platform.
-*‘TensorFlow’* is a highly efficient scientific computing framework, designed for machine learning (ML). It has two key advantages: (i) It can be run easily on heterogenous hardware, including CPUs and GPUs, with no or minimal code changes, and, (ii) It enables automatic differentiation, which is essential for deep learning, but also convenient for other optimization problems, such as iterative types of MRI reconstruction. 
-However, MRI reconstructions often rely on MRI-specific software, which does not typically interface easily with TensorFlow’s computing model. We developed *‘TensorFlow MRI’* to address this, by extending the TensorFlow ecosystem to seamlessly incorporate MRI-specific functionality.
[…]
