+++ actions-runner/_work/tensorflow-mri/tensorflow-mri/paper.bib
@@ -1,88 +1,0 @@
-
-@misc{fidgit,
-  author = {A. M. Smith and K. Thaney and M. Hahnel},
-  title = {Fidgit: An ungodly union of GitHub and Figshare},
-  year = {2020},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/arfon/fidgit}
-}
-
-@article{BakerDenoising,
-   author = {Baker, Rebecca R. and Muthurangu, Vivek and Rega, Marilena and Walsh, Stephen B. and Steeden, Jennifer A.},
-   title = {Rapid 2D 23Na MRI of the calf using a denoising convolutional neural network},
-   journal = {Magnetic Resonance Imaging},
-   volume = {110},
-   pages = {184-194},
-   ISSN = {0730-725X},
-   DOI = {https://doi.org/10.1016/j.mri.2024.04.027},
-   url = {https://www.sciencedirect.com/science/article/pii/S0730725X24001401},
-   year = {2024},
-   type = {Journal Article}
-}
-
-@article{BarnettNUFFT,
-   author = {Barnett, Alexander H and Magland, Jeremy and af Klinteberg, Ludvig},
-   title = {A parallel nonuniform fast Fourier transform library based on an “exponential of semicircle" kernel},
-   journal = {SIAM Journal on Scientific Computing},
-   volume = {41},
-   number = {5},
-   pages = {C479-C504},
-   ISSN = {1064-8275},
-   year = {2019},
-   type = {Journal Article}
-}
-
-@article{HansenGadgetron,
-   author = {Hansen, Michael Schacht and Sørensen, Thomas Sangild},
-   title = {Gadgetron: An open source framework for medical image reconstruction},
-   journal = {Magnetic Resonance in Medicine},
-   volume = {69},
-   number = {6},
-   pages = {1768-1776},
-   ISSN = {0740-3194},
-   DOI = {https://doi.org/10.1002/mrm.24389},
-   url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/mrm.24389},
-   year = {2013},
-   type = {Journal Article}
-}
-
-@article{JaubertHyperSlice,
-   author = {Jaubert, Olivier and Montalt-Tordera, Javier and Knight, Daniel and Arridge, Simon and Steeden, Jennifer and Muthurangu, Vivek},
-   title = {HyperSLICE: HyperBand optimized spiral for low-latency interactive cardiac examination},
-   journal = {Magnetic Resonance in Medicine},
-   volume = {91},
-   number = {1},
-   pages = {266-279},
-   ISSN = {0740-3194},
-   DOI = {https://doi.org/10.1002/mrm.29855},
-   url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/mrm.29855},
-   year = {2024},
-   type = {Journal Article}
-}
-
-@article{JaubertNaturalVideos,
-   author = {Jaubert, Olivier and Pascale, Michele and Montalt-Tordera, Javier and Akesson, Julius and Virsinskaite, Ruta and Knight, Daniel and Arridge, Simon and Steeden, Jennifer and Muthurangu, Vivek},
-   title = {Training deep learning based dynamic MR image reconstruction using open-source natural videos},
-   journal = {Scientific Reports},
-   volume = {14},
-   number = {1},
-   pages = {11774},
-   ISSN = {2045-2322},
-   DOI = {10.1038/s41598-024-62294-7},
-   url = {https://doi.org/10.1038/s41598-024-62294-7},
-   year = {2024},
-   type = {Journal Article}
-}
-
-@inproceedings{ShihNUFFT,
[…]
