### 2025-06-14 #8000 @Kevin-Mattheus-Moerman
@paquiteau, @mathieuboudreau, @bwheelz36, @cncastillo, @agahkarakuzu, @grlee77, @mfroeling, @62442katieb, @michaelberks, @matteomancini, @jonbmartin, @bwheelz36, @curtcorum  would you be interested in reviewing this submission for [JOSS (the Journal of Open Source Software](https://joss.theoj.org)? The work is titled _TensorFlow MRI: An Open-Source TensorFlow Library for MRI Applications_.

JOSS reviews take place on GitHub and focus on [the software :floppy_disk: ](https://github.com/mrphys/tensorflow-mri) as well as [a short paper :scroll: ](https://raw.githubusercontent.com/openjournals/joss-papers/joss.08000/joss.08000/10.21105.joss.08000.pdf). We typically aim for a 2-3 week review but we can be more flexible if needed. You can find more about our review process in our [review guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).  Here is an example of a review if you want to see what it is like: https://github.com/openjournals/joss-reviews/issues/7564.

If you are interested you can let me know by responding here.

Thanks!
