### 2025-06-15 #8410 @editorialbot
[REVIEW]: TensorFlow MRI: An Open-Source TensorFlow Library for MRI Applications
**Submitting author:** @jennifersteeden (<a href="http://orcid.org/0000-0002-9792-2022">Jennifer Steeden</a>)
**Repository:** https://github.com/mrphys/tensorflow-mri
**Branch with paper.md** (empty if default branch): master
**Version:** v0.23.1
