### 2025-07-19 mrphys/tensorflow-mri#51 @curtcorum
Joss review comments, concerns and suggestions.
@jmontalt, @jennifersteeden, @pascalemp

@cncastillo, @paquiteau, @Kevin-Mattheus-Moerman

Checklist: https://github.com/openjournals/joss-reviews/issues/8410#issuecomment-2978605575

First I concur with ALL the comments of reviewer @paquiteau: https://github.com/openjournals/joss-reviews/issues/8410#issuecomment-2976674787

Especially:
> I would recommend to move to pyproject.toml [setup](https://packaging.python.org/en/latest/tutorials/packaging-projects/); Using a setup.py is deprecated

Also of note:

[README.rst](https://github.com/mrphys/tensorflow-mri/blob/98c0fdec94a763b7feb49814cf50ffb0df4e1802/README.rst) is not up to date, latest released version is [v0.23.1](https://github.com/mrphys/tensorflow-mri/releases/tag/v0.23.1)

Ideally the JOSS paper references a specific release and potentially changes made in review, should this be v0.23.1? Upcoming v0.23.2?

It looks like there are a lot of changes between the current main head 98c0fdec94a763b7feb49814cf50ffb0df4e1802 and v0.23.1, was this a forced update or other variation? There is too much to track: https://github.com/mrphys/tensorflow-mri/compare/v0.23.1...master
