### 2025-07-19 mrphys/tensorflow-mri#51 @curtcorum
As an experiment I used copilot to generate a checklist:

# JOSS Submission Checklist for curtcorum/tensorflow-mri

Use this checklist to ensure your repository meets the JOSS guidelines before submission.

Some of it was probably incorrect based on expecting a README.md. I pruned it to items I agree with:

- [ ] **README.md** at repository root includes:
  - [ ] Project overview and scientific purpose **(some description, but unfocused)**
  - [ ] Statement of need (scientific problem addressed, target audience) **(some description, but unfocused)**
  - [ ] Installation instructions [**out of date setup.py**](https://github.com/openjournals/joss-reviews/issues/8410#issuecomment-2976528783)
  - [ ] Usage examples [**exist but not linked directly in README.md**](https://mrphys.github.io/tensorflow-mri/tutorials/recon/partialFourier/)
  - [ ] Links to documentation and related work [**no Docs for v.23.1**](https://mrphys.github.io/tensorflow-mri/)
  - [ ] Citation instructions (BibTeX/DOI) **Zenodo links not up to date for up to v0.23.1**
  - [ ] License and **authorship** information

- [ ] **API Documentation** (if applicable) [**need to update**](https://mrphys.github.io/tensorflow-mri/)
  - [ ] Generated docs (e.g., Sphinx for Python, JSDoc for JS/TS)
  - [ ] Usage and function/class descriptions

- [ ] **Changelog** (`CHANGELOG.md`) at root documents user-facing changes and **recent** releases

## Project Management
- [ ] **Contribution guidelines** (`CONTRIBUTING.md`) at root
  - [ ] Coding style and conventions
  - [ ] Instructions for submitting issues and pull requests
  - [ ] How to run tests
  - [ ] Reference to the Code of Conduct

- [ ] **Code of Conduct** (`CODE_OF_CONDUCT.md`) at root

- [ ] **Authors and affiliations** listed in README or `AUTHORS.md`

## Software Quality
- [ ] Consistent coding style (e.g., PEP8 for Python, ESLint for JS/TS)
- [ ] Linters and code formatting tools are used and documented
- [ ] Automated tests included and described
- [ ] Instructions for running tests provided
- [ ] Continuous integration configured (e.g., GitHub Actions)

## JOSS-specific
- [ ] License file (`LICENSE`) at root and mentioned in README
- [ ] Citation file (`CITATION.cff` or instructions in README)
- [ ] Statement of need clearly present in documentation
- [ ] Related work referenced and compared

## Other
- [ ] Project directories and files clearly organized
- [ ] Remove unnecessary template/boilerplate files
- [ ] Semantic versioning used and documented in changelog

---

_Adapted for JOSS review based on https://joss.theoj.org/papers/instructions_
