### 2025-07-19 mrphys/tensorflow-mri#51 @curtcorum
More copilot:

Based on the current semantic search, there is no direct evidence of PEP8 compliance or violations in the main codebase of curtcorum/tensorflow-mri. The results mainly show documentation, licenses, and configuration for submodules and GitHub Actions, not the core Python MRI code.

**What this means:**
- There is no clear indication that PEP8 style checks (e.g., via `flake8` or `pycodestyle`) are enforced for the Python source code in this repository.
- Without seeing Python source files, it is not possible to confirm or comment on specific style violations (such as indentation, line length, naming conventions, etc.).
- JOSS requires code to follow community best practices, and for Python this means PEP8 compliance.

**Recommendations:**
1. **Run a linter:** Use `flake8`, `pycodestyle`, or `black` on the core Python code and address any reported issues.
2. **Document Style:** Add a note in `CONTRIBUTING.md` that contributions must follow PEP8.
3. **CI Integration:** Add a GitHub Actions job that checks PEP8 compliance on each push/PR.
4. **Manual Review:** Spot-check for common issues: long lines, inconsistent indentation, naming conventions, and spacing.

Would you like instructions or a sample workflow for adding Python style checks to your repository?
