### 2025-07-19 mrphys/tensorflow-mri#51 @curtcorum
Malformed link in README.rst, possibly works in README.md?

`[![Watch the video](tools/assets/thumb.png)](https://vimeo.com/1054518675/e19c8abad3)
`
