### 2025-07-19 mrphys/tensorflow-mri#51 @curtcorum
With current pip based setup it is probably most appropriate to give or at least point to instructions for setting up a venv:

https://packaging.python.org/en/latest/guides/installing-using-pip-and-virtual-environments/
