### 2025-07-19 mrphys/tensorflow-mri#51 @curtcorum
Also supported (and tested) python versions. #50
