### 2025-07-19 mrphys/tensorflow-mri#51 @jmontalt
Dear @curtcorum ,

Thank you so much for your review and your valuable comments. I agree with virtually everything. Here are some of my first thoughts:

- `setup.py`: while I also prefer `pyproject.toml`, we will need to find a way to adapt the custom packaging workflow used to compile native C++ ops.
- Regarding Python versions: we need to upgrade the C++ ops to support modern Python versions. This is complicated by recent changes in how TensorFlow binds to custom ops, but in my opinion should take maximum priority.
- I believe some of the automated workflows (eg, the deployment of the latest docs) were not correctly triggered. The release process needs to be better documented or made more robust to human error.
- Some mistakes seem to have happened in recent releases and in the master branch, which caused the failed builds and explain the deviation between master and the latest release.

I will be in touch with my co-authors to find a way to address these and all other points. However, I expect this will take some time, not least because I work elsewhere now and my availability is limited. Is there a deadline?
