### 2025-07-19 mrphys/tensorflow-mri#51 @jmontalt
> Malformed link in README.rst, possibly works in README.md?
>
> `[![Watch the video](tools/assets/thumb.png)](https://vimeo.com/1054518675/e19c8abad3)
> `
>

Indeed this is Markdown syntax, not RestructuredText, which would explain the error. Thank you for pointing this out!

I would personally prefer to use .md files everywhere (more common, easier to write), but this has implications on the docs build workflow. Back then, I had to work around various limitations of Sphinx and associated extensions, but I suspect there may be better ways to do this nowadays. It might be a good idea to re-assess this soon.
