### 2025-07-19 mrphys/tensorflow-mri#51 @curtcorum
@jmontalt, @jennifersteeden, @pascalemp

@cncastillo, @paquiteau, @Kevin-Mattheus-Moerman

Review Summary:

Clearly tensorflow-mri has been and is a powerful package which would be a great addition to the general mri community. Unfortunately the package currently needs much updating, for consistency (between documentation and repository) and more testing for robustness in various linux and python environments.

Currently it does not seem generally installable or compatible without excessive effort, luck, or help from the developers.

The documentation in the repository is incomplete and it seems that previously functioning notebooks have been removed from the current repository. The previous documentation and examples are quite extensive. It is not clear where the source for this resides, and ideally is a part of the up to date repository.

I'd like to thank the developers for their significant past and current efforts, but in this reviewer's opinion, much work is needed to bring the project up to joss standards.
