### 2025-07-19 mrphys/tensorflow-mri#51 @curtcorum
@jmontalt I do not believe there is a specific deadline, editor @Kevin-Mattheus-Moerman should be able to comment.

Again, I think bringing tensorflow-mri up to joss standards would be a significant contribution to the mri open source community!
