### 2025-07-24 mrphys/tensorflow-mri#51 @curtcorum
@jmontalt, @jennifersteeden, @pascalemp

@cncastillo, @paquiteau, @Kevin-Mattheus-Moerman

Dear all, I believe the authors are still working/responding on the issues brought up by the reviewers.

https://github.com/mrphys/tensorflow-mri/issues/51#issuecomment-3092599622
> Dear [@curtcorum](https://github.com/curtcorum) ,
>
> Thank you so much for your review and your valuable comments. I agree with virtually everything. Here are some of my first thoughts:
>
>     * `setup.py`: while I also prefer `pyproject.toml`, we will need to find a way to adapt the custom packaging workflow used to compile native C++ ops.
>
>     * Regarding Python versions: we need to upgrade the C++ ops to support modern Python versions. This is complicated by recent changes in how TensorFlow binds to custom ops, but in my opinion should take maximum priority.
>
>     * I believe some of the automated workflows (eg, the deployment of the latest docs) were not correctly triggered. The release process needs to be better documented or made more robust to human error.
>
>     * Some mistakes seem to have happened in recent releases and in the master branch, which caused the failed builds and explain the deviation between master and the latest release.
>
>
> I will be in touch with my co-authors to find a way to address these and all other points. However, I expect this will take some time, not least because I work elsewhere now and my availability is limited. Is there a deadline?

https://github.com/mrphys/tensorflow-mri/issues/51#issuecomment-3092602145
> > Malformed link in README.rst, possibly works in README.md?
> > `[![Watch the video](tools/assets/thumb.png)](https://vimeo.com/1054518675/e19c8abad3) `
>
> Indeed this is Markdown syntax, not RestructuredText, which would explain the error. Thank you for pointing this out!
>
> I would personally prefer to use .md files everywhere (more common, easier to write), but this has implications on the docs build workflow. Back then, I had to work around various limitations of Sphinx and associated extensions, but I suspect there may be better ways to do this nowadays. It might be a good idea to re-assess this soon.

https://github.com/openjournals/joss-reviews/issues/8410#issuecomment-2976674787
