### 2025-07-24 #8410 @curtcorum
> [@curtcorum](https://github.com/curtcorum), [@paquiteau](https://github.com/paquiteau), [@cncastillo](https://github.com/cncastillo) can you check in with this review again please? Thanks!

Please see:

https://github.com/mrphys/tensorflow-mri/issues/51#issuecomment-3115235604
